If anything in name2id conversion fails then just return 0.
Do not set any errno. So if a non empty string is used in name2id
conversion and the return value is 0 then that would be an error.
Callers in most cases do not care and accept that a label may be
lost because of conversion failure.

Noticed by and OK tb@