Extended nexthop encoding support (RFC8950) for the RIB
This diff allows to establish IPv6 sessions and distribute IPv4 routes
with IPv6 nexthop over those sessions.

This is enough to use RFC 8950 on route reflectors or route servers.
It adds the support for IPv6 nexthops for IPv4 routes to the RDE / RIB.

The problem is that the FIB is not ready for that yet (the nexthop of a
kroute has to be of the same address family as the prefix) and fixing that
is not trivial. Because of this prefixes using a nexthop in a different
address family are not sent to the FIB.

OK tb@