Flip the switch on reject as-set from default no to yes.
BGP AS_PATH AS_SET are deprecated (or in the process to be).
In short AS_SET don't play nice with Route Origin Validation (ROV) and
with ASPA validation any AS_SET makes the path invalid and ineligible.
Right now there are some 200 routes left in the DFZ that use AS_SET so
the impact of this is minimal.

While there cleanup the code and remove some double negations and
NO defines.

OK sthen@ tb@, job@ agrees