Always set the peer holdtimer and min_holdtimer value
In init_peer and merge_peers initialize the holdtimer / min_holdtimer
of the peer from the global config if the peer value is unset.
This simplifies handling of the holdtimer in the message handling code.
Also this fixes a bug where the global min_holdtimer value was not
respected.
OK tb@