Fix RDE busy loop around poll because of peer_work_pending misreporting
The switch to ibufq introduced a bug with the imsg_pending tracking.
peer_imsg_flush() did not adjust imsg_pending so when a peer is reset
in the wrong moment imsg_pending becomes off and so peer_work_pending
would always return 1.

Lets simply walk the peertable in peer_work_pending. This is more work
but less error prone.

OK tb@