In pfkey_reply() ensure that any error message is properly discarded.
In rev 1.64 I tried to fix this but missed the ESRCH case and so an
ESRCH error would get stuck in the socket buffer and cause bgpd to
loop in the parent main event loop peeking at the same pfkey message
over and over again.

Bug report and ktrace provided by Tom Smyth.
OK tb@