Rewrite the find first empty spot code.
That code in ch_sub_insert() tries to be very clever but many tools are
not able to see through all the dependencies and trigger false positives.
Check the no empty condition first and else check if ins_g is still NULL.
There is no need to check the outcome of ffs() anymore and also the break
condition can only happen in this branch.

Reported by smatch via jsg@
CID 492341
OK tb@